00Information_Gathering/参考资料/information_gathering_beta-master/requirement.txt
requests
bs4
python-whois
python-nmap
